backend/requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
cycler==0.10.0
cymem==2.0.3
dateutils==0.6.8
dnspython==1.16.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
html2text==2020.1.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
murmurhash==1.0.2
nltk==3.5
nsetools==1.0.11
numpy==1.18.2
pandas==1.0.3
patsy==0.5.1
plac==1.1.3
preshed==3.0.2
PyJWT==1.7.1
pymongo[tls,srv]==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
regex==2020.5.7
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
spacy==2.2.4
srsly==1.0.2
statsmodels==0.11.1
thinc==7.4.0
tqdm==4.46.0
urllib3==1.25.9
waitress==1.4.3
wasabi==0.6.0
Werkzeug==1.0.1
